Minor fixes, documentation

+++ util/EncodingManager.java
@@ -1,11 +1,14 @@
 //Alexander Weaver
-//Last update: 5-13-2015 1:52am
+//Last update: 6-1-2015 4:33pm
 package Util;
 
 import java.math.BigInteger;
 
+//Manages encoding/padding of strings
 public class EncodingManager {
     
+    //Takes a string in ASCII
+    //Converts each character to its hex code, and returns the concatenated hex string
     public String textToHexASCII(String text) {
         char[] chars = text.toCharArray();
         StringBuilder hex = new StringBuilder();
@@ -19,16 +22,21 @@ public class EncodingManager {
         return hex.toString();
     }
     
+    //Converts a hex string to a BigInteger format
     public BigInteger hexToBigInteger(String hex) {
         BigInteger result = new BigInteger(hex, 16);
         return result;
     }
     
+    //Converts a BigInteger to a hex string format
     public String bigIntegerToHex(BigInteger value) {
         String result = value.toString(16);
         return result;
     }
     
+    //Takes a string of hexadecimal values
+    //Groups every two values and converts them to their corresponding ASCII character
+    //returns a string of concatenated characters
     public String hexToTextASCII(String hex) {
         StringBuilder output = new StringBuilder("");
         for (int i = 0; i < hex.length(); i += 2) {
@@ -38,7 +46,10 @@ public class EncodingManager {
         return output.toString();
 }
     
-    public String valuesToTextASCII(short[] values) {
+    //Takes a byte array of numerical values
+    //Converts each byte to its corresponding ASCII character
+    //Returns a string of concatenated characters
+    public String valuesToTextASCII(byte[] values) {
         int length = values.length;
         StringBuilder sb = new StringBuilder();
         char current;
@@ -52,6 +63,9 @@ public class EncodingManager {
         return sb.toString();
     }
     
+    //Takes an int array of numerical values
+    //Converts each int to its corresponding ASCII character
+    //Returns a string of concatenated characters
     public String valuesToTextASCII(int[] values) {
         int length = values.length;
         StringBuilder sb = new StringBuilder();
@@ -66,6 +80,9 @@ public class EncodingManager {
         return sb.toString();
     }
     
+    //Takes a string and an integer section length
+    //Separates the string into smaller sub strings of length sectionLength
+    //Returns an array of all substrings
     public String[] separate(String text, int sectionLength) {
         int length = text.length();
         if(length == 0) {
@@ -82,6 +99,8 @@ public class EncodingManager {
         return separatedText;
     }
     
+    //Takes an array of strings
+    //Joins them into a single string
     public String joinStrings(String[] strings) {
         int length = strings.length;
         String joined = "";
